# backend/src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FoodFlow API",
    "version": "1.0.0",
    "description": "API para o sistema FoodFlow, um SaaS para gestão de restaurantes. Inclui autenticação, gestão de restaurantes, usuários, mesas, cardápio, sessões, pedidos, relatórios, reservas e personalização de estilo."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor local"
    },
    {
      "url": "https://foodflow-api.render.com",
      "description": "Servidor de produção (Render)"
    }
  ],
...
  "paths": {
...
    "/users/{restaurantId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Lista usuários de um restaurante",
        "description": "Retorna a lista de usuários (garçons, cozinheiros, etc.) associados a um restaurante. Requer JWT.",
        "security": [
          {
            "bearerauth": []
          }
        ],
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID do restaurante"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuários",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  }
                },
                "example": [
                  {
                    "id": "user_1",
                    "email": "garcom@restaurante.com",
                    "name": "Ana Souza",
                    "role_id": "3f3aed51-f815-40dc-a372-a31de658319f",
                    "restaurant_id": "rest_1"
                  }
                ]
              }
            }
          },
          "400": {
            "description": "Erro ao listar",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Restaurante não encontrado"
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Atualiza um usuário",
        "description": "Atualiza o papel, restaurante ou status de um usuário. Requer JWT.",
        "security": [
          {
            "bearerauth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID do usuário"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserInput"
              },
              "example": {
                "role_id": "3f3aed51-f815-40dc-a372-a31de658319f",
                "restaurant_id": "rest_1",
                "is_active": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                },
                "example": {
                  "id": "user_1",
                  "email": "garcom@restaurante.com",
                  "name": "Ana Souza",
                  "role_id": "3f3aed51-f815-40dc-a372-a31de658319f",
                  "restaurant_id": "rest_1"
                }
              }
            }
          },
          "400": {
            "description": "Erro ao atualizar",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Usuário não encontrado"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Exclui um usuário",
        "description": "Remove um usuário do sistema. Requer JWT.",
        "security": [
          {
            "bearerauth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID do usuário"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Usuário excluído"
          },
          "400": {
            "description": "Erro ao excluir",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Usuário não encontrado"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerauth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtido via /auth/login"
      }
    },
    "schemas": {
...
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "ID do usuário"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "E-mail do usuário"
          },
          "name": {
            "type": "string",
            "description": "Nome do usuário"
          },
          "role_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID do papel"
          },
          "restaurant_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID do restaurante",
            "nullable": true
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role_id"
        ]
      },
...
      "UpdateUserInput": {
        "type": "object",
        "properties": {
          "role_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID do papel",
            "nullable": true
          },
          "restaurant_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID do restaurante",
            "nullable": true
          },
          "is_active": {
            "type": "boolean",
            "description": "Status do usuário",
            "nullable": true
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Mensagem de erro",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "description": "Mensagem de erro específica"
                },
                "path": {
                  "type": "string",
                  "description": "Campo com erro"
                },
                "location": {
                  "type": "string",
                  "description": "Local do erro (ex.: body, query)"
                }
              }
            },
            "description": "Lista de erros detalhados",
            "nullable": true
          }
        }
      }
    }
  }
}